+++ vast_pipeline/pipeline/utils.py
@@ -14,14 +14,15 @@ import astropy.units as u
 import dask
 import dask.dataframe as dd
 
-from typing import List, Optional, Dict, Tuple
+from typing import Any, List, Optional, Dict, Tuple, Union
 from astropy.io import fits
 from astropy.coordinates import SkyCoord, Angle
 from django.conf import settings
 from django.contrib.auth.models import User
 from psutil import cpu_count
-from itertools import chain, combinations
+from itertools import chain
 
+from vast_pipeline.image.main import FitsImage, SelavyImage
 from vast_pipeline.utils.utils import (
     eq_to_cart, StopWatch, optimize_ints, optimize_floats
 )
@@ -48,13 +49,13 @@ def get_create_skyreg(image: Image) -> SkyRegion:
     # In the calculations below, it is assumed the image has square
     # pixels (this pipeline has been designed for ASKAP images, so it
     # should always be square). It will likely give wrong results if not
-    skyr = SkyRegion.objects.filter(
+    skyregions = SkyRegion.objects.filter(
         centre_ra=image.ra,
         centre_dec=image.dec,
         xtr_radius=image.fov_bmin
     )
-    if skyr:
-        skyr = skyr.get()
+    if skyregions:
+        skyr = skyregions.get()
         logger.info('Found sky region %s', skyr)
     else:
         x, y, z = eq_to_cart(image.ra, image.dec)
@@ -74,7 +75,7 @@ def get_create_skyreg(image: Image) -> SkyRegion:
     return skyr
 
 
-def get_create_img_band(image: Image) -> Band:
+def get_create_img_band(image: FitsImage) -> Band:
     '''
     Return the existing Band row for the given FitsImage.
     An image is considered to belong to a band if its frequency is within some
@@ -106,23 +107,23 @@ def get_create_img_band(image: Image) -> Band:
     return band
 
 
-def get_create_img(band_id: int, image: Image) -> Tuple[Image, bool]:
+def get_create_img(band_id: int, image: SelavyImage) -> Tuple[Image, bool]:
     """
     Function to fetch or create the Image and Sky Region objects for an image.
 
     Args:
         band_id: The integer database id value of the frequency band of the
             image.
-        image: The image Django ORM object.
+        image: The image object.
 
     Returns:
-        The resulting image django ORM object, and a bool value denoting if the image already existed in the
-        database.
+        The resulting image django ORM object, and a bool value denoting if the image
+        already existed in the database.
     """
-    img = Image.objects.filter(name__exact=image.name)
-    exists = img.exists()
+    images = Image.objects.filter(name__exact=image.name)
+    exists = images.exists()
     if exists:
-        img = img.get()
+        img: Image = images.get()
         # Add background path if not originally provided
         if image.background_path and not img.background_path:
             img.background_path = image.background_path
@@ -160,7 +161,7 @@ def get_create_img(band_id: int, image: Image) -> Tuple[Image, bool]:
 
 
 def get_create_p_run(
-    name: str, path: str, description: str=None, user: User=None
+    name: str, path: str, description: str = None, user: User = None
 ) -> Tuple[Run, bool]:
     '''
     Get or create a pipeline run in db, return the run django object and
@@ -248,8 +249,6 @@ def remove_duplicate_measurements(
         'Using duplicate crossmatch radius of %.2f arcsec.', dup_lim.arcsec
     )
 
-    min_source = sources_df['source'].min()
-
     # sort by the distance from the image centre so we know
     # that the first source is always the one to keep
     sources_df = sources_df.sort_values(by='dist_from_centre')
@@ -261,7 +260,7 @@ def remove_duplicate_measurements(
     )
 
     # perform search around sky to get all self matches
-    idxc, idxcatalog, d2d_around, _ = sources_sc.search_around_sky(
+    idxc, idxcatalog, *_ = sources_sc.search_around_sky(
         sources_sc, dup_lim
     )
 
@@ -495,7 +494,7 @@ def add_new_one_to_many_relations(
             out = row[related_col]
             out.append(row[source_col])
         else:
-            out = [row[source_col],]
+            out = [row[source_col], ]
 
     else:  # the original case to append all the new ids.
         source_ids = source_ids.loc[row[source_col]].iloc[0]
@@ -552,7 +551,7 @@ def cross_join(left: pd.DataFrame, right: pd.DataFrame) -> pd.DataFrame:
 
 
 def get_eta_metric(
-    row: Dict[str, float], df: pd.DataFrame, peak: bool=False
+    row: Dict[str, float], df: pd.DataFrame, peak: bool = False
 ) -> float:
     '''
     Calculates the eta variability metric of a source.
@@ -606,7 +605,7 @@ def groupby_funcs(df: pd.DataFrame) -> pd.Series:
     d['n_meas_sel'] = d['n_meas'] - d['n_meas_forced']
     d['n_sibl'] = df['has_siblings'].sum()
     if d['n_meas_forced'] > 0:
-        non_forced_sel = df['forced'] != True
+        non_forced_sel = ~df['forced']
         d['wavg_ra'] = (
             df.loc[non_forced_sel, 'interim_ew'].sum() /
             df.loc[non_forced_sel, 'weight_ew'].sum()
@@ -804,7 +803,7 @@ def get_rms_noise_image_values(rms_path: str) -> Tuple[float, float, float]:
     return med_val, min_val, max_val
 
 
-def get_image_list_diff(row: pd.Series) -> List[str]:
+def get_image_list_diff(row: pd.Series) -> Union[List[str], int]:
     """
     Calculate the difference between the ideal coverage image list of a source
     and the actual observed image list. Also checks whether an epoch does in
@@ -841,7 +840,7 @@ def get_image_list_diff(row: pd.Series) -> List[str]:
     ]
 
     if not out:
-        out = -1
+        return -1
 
     return out
 
@@ -860,7 +859,7 @@ def get_names_and_epochs(grp: pd.DataFrame) -> pd.Series:
         image names, epochs and datetimes.
     """
     d = {}
-    d['skyreg_img_epoch_list'] = [[[x,], y, z] for x,y,z in zip(
+    d['skyreg_img_epoch_list'] = [[[x, ], y, z] for x, y, z in zip(
         grp['name'].values.tolist(),
         grp['epoch'].values.tolist(),
         grp['datetime'].values.tolist()
@@ -1027,12 +1026,12 @@ def get_src_skyregion_merged_df(
     # annoyingly epoch needs to be not a list to drop duplicates
     # but then we need to sum the epochs into a list
     src_skyrg_df['skyreg_epoch'] = src_skyrg_df['skyreg_epoch'].apply(
-        lambda x: [x,]
+        lambda x: [x, ]
     )
 
     src_skyrg_df = (
         src_skyrg_df.groupby('source')
-        .sum(numeric_only=False) # sum because we need to preserve order
+        .sum(numeric_only=False)  # sum because we need to preserve order
     )
 
     # merge into main df and compare the images
@@ -1164,7 +1163,7 @@ def group_skyregions(df: pd.DataFrame) -> pd.DataFrame:
         axis=1
     )
 
-    skyreg_groups = {}
+    skyreg_groups: Dict[int, List[Any]] = {}
 
     master_done = []  # keep track of all checked ids in master done
 
@@ -1214,7 +1213,7 @@ def group_skyregions(df: pd.DataFrame) -> pd.DataFrame:
     skyreg_group_ids = {}
     for i in skyreg_groups:
         for j in skyreg_groups[i]:
-            skyreg_group_ids[j]=i
+            skyreg_group_ids[j] = i
 
     skyreg_group_ids = pd.DataFrame.from_dict(
         skyreg_group_ids, orient='index'
@@ -1350,8 +1349,8 @@ def create_measurements_arrow_file(p_run: Run) -> None:
     local = pa.fs.LocalFileSystem()
 
     with local.open_output_stream(outname) as file:
-       with pa.RecordBatchFileWriter(file, measurements.schema) as writer:
-          writer.write_table(measurements)
+        with pa.RecordBatchFileWriter(file, measurements.schema) as writer:
+            writer.write_table(measurements)
 
 
 def create_measurement_pairs_arrow_file(p_run: Run) -> None:
@@ -1387,183 +1386,8 @@ def create_measurement_pairs_arrow_file(p_run: Run) -> None:
     local = pa.fs.LocalFileSystem()
 
     with local.open_output_stream(outname) as file:
-       with pa.RecordBatchFileWriter(file, measurement_pairs_df.schema) as writer:
-          writer.write_table(measurement_pairs_df)
-
-
-def calculate_vs_metric(
-    flux_a: float, flux_b: float, flux_err_a: float, flux_err_b: float
-) -> float:
-    """Calculate the Vs variability metric which is the t-statistic that the provided
-    fluxes are variable. See Section 5 of Mooley et al. (2016) for details,
-    DOI: 10.3847/0004-637X/818/2/105.
-
-    Args:
-        flux_a (float): flux value "A".
-        flux_b (float): flux value "B".
-        flux_err_a (float): error of `flux_a`.
-        flux_err_b (float): error of `flux_b`.
-
-    Returns:
-        float: the Vs metric for flux values "A" and "B".
-    """
-    return (flux_a - flux_b) / np.hypot(flux_err_a, flux_err_b)
-
-
-def calculate_m_metric(flux_a: float, flux_b: float) -> float:
-    """Calculate the m variability metric which is the modulation index between two fluxes.
-    This is proportional to the fractional variability.
-    See Section 5 of Mooley et al. (2016) for details, DOI: 10.3847/0004-637X/818/2/105.
-
-    Args:
-        flux_a (float): flux value "A".
-        flux_b (float): flux value "B".
-
-    Returns:
-        float: the m metric for flux values "A" and "B".
-    """
-    return 2 * ((flux_a - flux_b) / (flux_a + flux_b))
-
-
-def calculate_measurement_pair_metrics(df: pd.DataFrame) -> pd.DataFrame:
-    """Generate a DataFrame of measurement pairs and their 2-epoch variability metrics
-    from a DataFrame of measurements. For more information on the variability metrics, see
-    Section 5 of Mooley et al. (2016), DOI: 10.3847/0004-637X/818/2/105.
-
-    Args:
-        df (pd.DataFrame): Input measurements. Must contain columns: id, source, flux_int,
-            flux_int_err, flux_peak, flux_peak_err, has_siblings.
-
-    Returns:
-        pd.DataFrame: Measurement pairs and 2-epoch metrics. Will contain columns:
-            source - the source ID
-            id_a, id_b - the measurement IDs
-            flux_int_a, flux_int_b - measurement integrated fluxes in mJy
-            flux_int_err_a, flux_int_err_b - measurement integrated flux errors in mJy
-            flux_peak_a, flux_peak_b - measurement peak fluxes in mJy/beam
-            flux_peak_err_a, flux_peak_err_b - measurement peak flux errors in mJy/beam
-            vs_peak, vs_int - variability t-statistic
-            m_peak, m_int - variability modulation index
-    """
-    n_cpu = cpu_count() - 1
-
-    """Create a DataFrame containing all measurement ID combinations per source.
-    Resultant DataFrame will have a MultiIndex(["source", RangeIndex]) where "source" is
-    the source ID and RangeIndex is an unnamed temporary ID for each measurement pair,
-    unique only together with source.
-    DataFrame will have columns [0, 1], each containing a measurement ID. e.g.
-                       0      1
-        source
-        1       0      1   9284
-                1      1  17597
-                2      1  26984
-                3   9284  17597
-                4   9284  26984
-        ...          ...    ...
-        11105   2  11845  19961
-        11124   0   3573  12929
-                1   3573  21994
-                2  12929  21994
-        11128   0   6216  23534
-    """
-    measurement_combinations = (
-        dd.from_pandas(df, n_cpu)
-        .groupby("source")["id"]
-        .apply(
-            lambda x: pd.DataFrame(list(combinations(x, 2))
-        ), meta={0: "i", 1: "i"},)
-        .compute(num_workers=n_cpu, scheduler="processes")
-    )
-
-    """Drop the RangeIndex from the MultiIndex as it isn't required and rename the columns.
-    Example resultant DataFrame:
-               source   id_a   id_b
-        0           1      1   9284
-        1           1      1  17597
-        2           1      1  26984
-        3           1   9284  17597
-        4           1   9284  26984
-        ...       ...    ...    ...
-        33640   11105  11845  19961
-        33641   11124   3573  12929
-        33642   11124   3573  21994
-        33643   11124  12929  21994
-        33644   11128   6216  23534
-    Where source is the source ID, id_a and id_b are measurement IDs.
-    """
-    measurement_combinations = measurement_combinations.reset_index(
-        level=1, drop=True
-    ).rename(columns={0: "id_a", 1: "id_b"}).astype(int).reset_index()
-
-    # Dask has a tendency to swap which order the measurement pairs are
-    # defined in, even if the dataframe is pre-sorted. We want the pairs to be
-    # in date order (a < b) so the code below corrects any that are not.
-    measurement_combinations = measurement_combinations.join(
-        df[['source', 'id', 'datetime']].set_index(['source', 'id']),
-        on=['source', 'id_a'],
-    )
-
-    measurement_combinations = measurement_combinations.join(
-        df[['source', 'id', 'datetime']].set_index(['source', 'id']),
-        on=['source', 'id_b'], lsuffix='_a', rsuffix='_b'
-    )
-
-    to_correct_mask = (
-        measurement_combinations['datetime_a']
-        > measurement_combinations['datetime_b']
-    )
-
-    if np.any(to_correct_mask):
-        logger.debug('Correcting measurement pairs order')
-        (
-            measurement_combinations.loc[to_correct_mask, 'id_a'],
-            measurement_combinations.loc[to_correct_mask, 'id_b']
-        ) = np.array([
-            measurement_combinations.loc[to_correct_mask, 'id_b'].values,
-            measurement_combinations.loc[to_correct_mask, 'id_a'].values
-        ])
-
-    measurement_combinations = measurement_combinations.drop(
-        ['datetime_a', 'datetime_b'], axis=1
-    )
-
-    # add the measurement fluxes and errors
-    association_fluxes = df.set_index(["source", "id"])[
-        ["flux_int", "flux_int_err", "flux_peak", "flux_peak_err", "image"]
-    ].rename(columns={"image": "image_name"})
-    measurement_combinations = measurement_combinations.join(
-        association_fluxes,
-        on=["source", "id_a"],
-    ).join(
-        association_fluxes,
-        on=["source", "id_b"],
-        lsuffix="_a",
-        rsuffix="_b",
-    )
-
-    # calculate 2-epoch metrics
-    measurement_combinations["vs_peak"] = calculate_vs_metric(
-        measurement_combinations.flux_peak_a,
-        measurement_combinations.flux_peak_b,
-        measurement_combinations.flux_peak_err_a,
-        measurement_combinations.flux_peak_err_b,
-    )
-    measurement_combinations["vs_int"] = calculate_vs_metric(
-        measurement_combinations.flux_int_a,
-        measurement_combinations.flux_int_b,
-        measurement_combinations.flux_int_err_a,
-        measurement_combinations.flux_int_err_b,
-    )
-    measurement_combinations["m_peak"] = calculate_m_metric(
-        measurement_combinations.flux_peak_a,
-        measurement_combinations.flux_peak_b,
-    )
-    measurement_combinations["m_int"] = calculate_m_metric(
-        measurement_combinations.flux_int_a,
-        measurement_combinations.flux_int_b,
-    )
-
-    return measurement_combinations
+        with pa.RecordBatchFileWriter(file, measurement_pairs_df.schema) as writer:
+            writer.write_table(measurement_pairs_df)
 
 
 def backup_parquets(p_run_path: str) -> None:
@@ -1765,7 +1589,7 @@ def reconstruct_associtaion_dfs(
 
     # Create the unique skyc1_srcs dataframe.
     skyc1_srcs = (
-        sources_df[sources_df['forced'] == False]
+        sources_df[~sources_df['forced']]
         .sort_values(by='id')
         .drop('related', axis=1)
         .drop_duplicates('source')
@@ -1775,7 +1599,7 @@ def reconstruct_associtaion_dfs(
     # which does not have the relation information)
     skyc1_srcs = skyc1_srcs.merge(
         prev_relations, how='left', left_on='source', right_index=True
-)
+    )
 
     # Need to break the pointer relationship between the related sources (
     # deep=True copy does not truly copy mutable type objects)